接受Qt示例的修改，去除其它内容，cmake 启用 MOC
