+++ CMakeLists.txt
@@ -6,31 +6,21 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 set(EXECUTABLE_OUTPUT_PATH ${PROJECT_SOURCE_DIR}/build/${CMAKE_BUILD_TYPE}/bin)
 
-include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include)
+set(CMAKE_AUTOMOC ON)
 
-option(USE_QT "Build Qt GUI example(s) instead of console" OFF)
+file(GLOB EXAMPLE_SOURCES "${CMAKE_CURRENT_SOURCE_DIR}/examples/*.cpp")
 
-if(USE_QT)
-    message(STATUS "Building Qt GUI example...")
+find_package(Qt5 COMPONENTS Core Widgets REQUIRED)
 
-    set(CMAKE_AUTOMOC ON)
-    find_package(Qt5 COMPONENTS Core Widgets REQUIRED)
+foreach(EXAMPLE_SOURCE IN ITEMS ${EXAMPLE_SOURCES})
+    get_filename_component(EXAMPLE_NAME ${EXAMPLE_SOURCE} NAME_WE)
+    set(EXAMPLE_EXECUTABLE example-${EXAMPLE_NAME})
 
-    file(GLOB QT_SOURCES "${CMAKE_CURRENT_SOURCE_DIR}/examples/qt-gui/*.cpp")
-    foreach(SRC ${QT_SOURCES})
-        get_filename_component(NAME ${SRC} NAME_WE)
-        add_executable(example-qt-${NAME} ${SRC})
-        target_link_libraries(example-qt-${NAME} Qt5::Core Qt5::Widgets)
-    endforeach()
+    add_executable(${EXAMPLE_EXECUTABLE} ${EXAMPLE_SOURCE})
+    
+    target_link_libraries(${EXAMPLE_EXECUTABLE} Qt5::Core Qt5::Widgets)
 
-else()
-    message(STATUS "Building Console example...")
-
-    file(GLOB CONSOLE_SOURCES "${CMAKE_CURRENT_SOURCE_DIR}/examples/console/*.cpp")
-    foreach(SRC ${CONSOLE_SOURCES})
-        get_filename_component(NAME ${SRC} NAME_WE)
-        add_executable(example-console-${NAME} ${SRC})
-    endforeach()
-
-endif()
-
+    target_include_directories(${EXAMPLE_EXECUTABLE}
+        PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/include
+    )
+endforeach()
